+++ sshd-core/src/main/java/org/apache/sshd/server/command/CommandDirectOutputStreamAware.java
@@ -17,16 +17,21 @@
  * under the License.
  */
 
-package org.apache.sshd.sftp.server;
+package org.apache.sshd.server.command;
 
-import org.apache.sshd.common.util.threads.ExecutorServiceCarrier;
+import java.io.OutputStream;
 
 /**
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
-public interface SftpSubsystemConfigurator
-        extends ExecutorServiceCarrier, SftpFileSystemAccessorProvider,
-        SftpUnsupportedAttributePolicyProvider, SftpErrorStatusDataHandlerProvider,
-        SftpErrorDataChannelReceiverProvider {
-    // Nothing extra
+@FunctionalInterface
+public interface CommandDirectOutputStreamAware {
+
+    /**
+     * Set the output stream that can be used by the shell to write its output.
+     *
+     * @param out The {@link OutputStream} used by the shell to write its output
+     */
+    void setOutputStream(OutputStream out);
+
 }